requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
Flask-Talisman==1.1.0 
web3==6.18.0
flask-pymongo==2.3.0